requirements.txt
# Core display libraries
luma.oled>=3.9.0  # SSD1306 OLED display driver
RPi.GPIO>=0.7.0  # GPIO access
Pillow>=9.0.0    # Image processing

# System monitoring
psutil>=5.9.0    # System statistics
docker>=6.0.0    # Docker container monitoring

# Configuration handling
pyyaml>=6.0     # YAML configuration file support

# Testing and documentation
pytest>=7.0.0   # Unit testing framework
sphinx>=5.0.0   # Documentation generator

# Development tools
black>=23.0.0   # Code formatter
isort>=5.10.0   # Import sorter
flake8>=6.0.0   # Linter

# Optional: Adafruit libraries as alternatives
# adafruit-circuitpython-ssd1306>=2.12.0
# adafruit-circuitpython-display-text>=2.22.0
